requirements.txt
# Install dependencies using pip install -r requirements.txt
pillow
pytesseract